requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-ckeditor==6.5.1
django-embed-video==1.4.5
django-heroku==0.3.1
django-js-asset==2.0.0
django-multiselectfield==0.1.12
gunicorn==20.1.0
humanize==4.3.0
idna==3.4
Pillow==9.2.0
psycopg2==2.9.3
python-decouple==3.6
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
